Функциональный объект для модуля
Close #296
